+++ internal/object/wire.go
@@ -4,7 +4,9 @@ import (
 	"errors"
 	"fmt"
 	"io"
+	"strconv"
 
+	iprotobuf "github.com/nspcc-dev/neofs-node/internal/protobuf"
 	"github.com/nspcc-dev/neofs-sdk-go/object"
 	protoobject "github.com/nspcc-dev/neofs-sdk-go/proto/object"
 	"github.com/nspcc-dev/neofs-sdk-go/proto/refs"
@@ -12,20 +14,13 @@ import (
 	"google.golang.org/protobuf/proto"
 )
 
-// Protobuf field numbers for object message.
-const (
-	_ = iota
-	fieldObjectID
-	fieldObjectSignature
-	fieldObjectHeader
-	fieldObjectPayload
-)
+var errEmptyData = errors.New("empty data")
 
 // WriteWithoutPayload writes the object header to the given writer without the payload.
 func WriteWithoutPayload(w io.Writer, obj object.Object) error {
 	header := obj.CutPayload().Marshal()
 	if obj.PayloadSize() != 0 {
-		header = protowire.AppendTag(header, fieldObjectPayload, protowire.BytesType)
+		header = protowire.AppendTag(header, protoobject.FieldObjectPayload, protowire.BytesType)
 		header = protowire.AppendVarint(header, obj.PayloadSize())
 	}
 	_, err := w.Write(header)
@@ -41,7 +36,7 @@ func ExtractHeaderAndPayload(data []byte) (*object.Object, []byte, error) {
 	)
 
 	if len(data) == 0 {
-		return nil, nil, fmt.Errorf("empty data")
+		return nil, nil, errEmptyData
 	}
 
 	for offset < len(data) {
@@ -55,7 +50,7 @@ func ExtractHeaderAndPayload(data []byte) (*object.Object, []byte, error) {
 			return nil, nil, fmt.Errorf("unexpected wire type: %v", typ)
 		}
 
-		if num == fieldObjectPayload {
+		if num == protoobject.FieldObjectPayload {
 			_, n = protowire.ConsumeVarint(data[offset:])
 			if err := protowire.ParseError(n); err != nil {
 				return nil, nil, fmt.Errorf("invalid varint at offset %d: %w", offset, err)
@@ -70,17 +65,17 @@ func ExtractHeaderAndPayload(data []byte) (*object.Object, []byte, error) {
 		offset += n
 
 		switch num {
-		case fieldObjectID:
+		case protoobject.FieldObjectID:
 			obj.ObjectId = new(refs.ObjectID)
 			if err := proto.Unmarshal(val, obj.ObjectId); err != nil {
 				return nil, nil, fmt.Errorf("unmarshal object ID: %w", err)
 			}
-		case fieldObjectSignature:
+		case protoobject.FieldObjectSignature:
 			obj.Signature = new(refs.Signature)
 			if err := proto.Unmarshal(val, obj.Signature); err != nil {
 				return nil, nil, fmt.Errorf("unmarshal object signature: %w", err)
 			}
-		case fieldObjectHeader:
+		case protoobject.FieldObjectHeader:
 			obj.Header = new(protoobject.Header)
 			if err := proto.Unmarshal(val, obj.Header); err != nil {
 				return nil, nil, fmt.Errorf("unmarshal object header: %w", err)
@@ -107,4 +102,146 @@ func ReadHeaderPrefix(r io.Reader) (*object.Object, []byte, error) {
 		return nil, nil, err
 	}
 	return ExtractHeaderAndPayload(buf[:n])
+}
+
+// GetNonPayloadFieldBounds seeks ID, signature and header in object message and
+// parses their boundaries.
+//
+// If buf is empty, GetNonPayloadFieldBounds returns an error.
+//
+// If any field is missing, no error is returned.
+//
+// Message should have ascending field order, otherwise error returns.
+func GetNonPayloadFieldBounds(buf []byte) (iprotobuf.FieldBounds, iprotobuf.FieldBounds, iprotobuf.FieldBounds, error) {
+	var idf, sigf, hdrf iprotobuf.FieldBounds
+	if len(buf) == 0 {
+		return idf, sigf, hdrf, errEmptyData
+	}
+
+	var off int
+	var prevNum protowire.Number
+loop:
+	for {
+		num, typ, n, err := iprotobuf.ParseTag(buf[off:])
+		if err != nil {
+			return idf, sigf, hdrf, err
+		}
+
+		if num > protoobject.FieldObjectHeader {
+			break
+		}
+		if num < prevNum {
+			return idf, sigf, hdrf, iprotobuf.NewUnorderedFieldsError(prevNum, num)
+		}
+		if num == prevNum {
+			return idf, sigf, hdrf, iprotobuf.NewRepeatedFieldError(num)
+		}
+		prevNum = num
+
+		f, err := iprotobuf.ParseLENFieldBounds(buf, off, n, num, typ)
+		if err != nil {
+			return idf, sigf, hdrf, err
+		}
+
+		switch num {
+		case protoobject.FieldObjectID:
+			idf = f
+		case protoobject.FieldObjectSignature:
+			sigf = f
+		case protoobject.FieldObjectHeader:
+			hdrf = f
+			break loop
+		default:
+			panic("unreachable with num " + strconv.Itoa(int(num)))
+		}
+
+		off = f.To
+
+		if off == len(buf) {
+			break
+		}
+	}
+
+	return idf, sigf, hdrf, nil
+}
+
+// GetParentNonPayloadFieldBounds seeks parent's ID, signature and header in child
+// object message and parses their boundaries.
+//
+// If buf is empty, GetParentNonPayloadFieldBounds returns an error.
+//
+// If any field is missing, no error is returned.
+//
+// Message should have ascending field order, otherwise error returns.
+func GetParentNonPayloadFieldBounds(buf []byte) (iprotobuf.FieldBounds, iprotobuf.FieldBounds, iprotobuf.FieldBounds, error) {
+	var idf, sigf, hdrf iprotobuf.FieldBounds
+	if len(buf) == 0 {
+		return idf, sigf, hdrf, errEmptyData
+	}
+
+	rootHdrf, err := iprotobuf.GetLENFieldBounds(buf, protoobject.FieldObjectHeader)
+	if err != nil {
+		return idf, sigf, hdrf, err
+	}
+
+	if rootHdrf.IsMissing() {
+		return idf, sigf, hdrf, nil
+	}
+
+	splitf, err := iprotobuf.GetLENFieldBounds(buf[rootHdrf.ValueFrom:rootHdrf.To], protoobject.FieldHeaderSplit)
+	if err != nil {
+		return idf, sigf, hdrf, err
+	}
+
+	if splitf.IsMissing() {
+		return idf, sigf, hdrf, nil
+	}
+
+	buf = buf[:rootHdrf.ValueFrom+splitf.To]
+	off := rootHdrf.ValueFrom + splitf.ValueFrom
+	var prevNum protowire.Number
+loop:
+	for {
+		num, typ, n, err := iprotobuf.ParseTag(buf[off:])
+		if err != nil {
+			return idf, sigf, hdrf, err
+		}
+
+		if num > protoobject.FieldHeaderSplitParentHeader {
+			break
+		}
+		if num < prevNum {
+			return idf, sigf, hdrf, iprotobuf.NewUnorderedFieldsError(prevNum, num)
+		}
+		if num == prevNum {
+			return idf, sigf, hdrf, iprotobuf.NewRepeatedFieldError(num)
+		}
+		prevNum = num
+
+		f, err := iprotobuf.ParseLENFieldBounds(buf, off, n, num, typ)
+		if err != nil {
+			return idf, sigf, hdrf, err
+		}
+
+		switch num {
+		case protoobject.FieldHeaderSplitParent:
+			idf = f
+		case protoobject.FieldHeaderSplitPrevious:
+		case protoobject.FieldHeaderSplitParentSignature:
+			sigf = f
+		case protoobject.FieldHeaderSplitParentHeader:
+			hdrf = f
+			break loop
+		default:
+			panic("unreachable with num " + strconv.Itoa(int(num)))
+		}
+
+		off = f.To
+
+		if off == len(buf) {
+			break
+		}
+	}
+
+	return idf, sigf, hdrf, nil
 }