+++ pkg/services/object/put/proto.go
@@ -7,31 +7,12 @@ import (
 
 	neofscrypto "github.com/nspcc-dev/neofs-sdk-go/crypto"
 	"github.com/nspcc-dev/neofs-sdk-go/object"
+	protoobject "github.com/nspcc-dev/neofs-sdk-go/proto/object"
+	protorefs "github.com/nspcc-dev/neofs-sdk-go/proto/refs"
 	"google.golang.org/protobuf/encoding/protowire"
 )
 
 // TODO: share common code with replication service code
-
-// Signature message fields.
-const (
-	_ = iota
-	fieldNumSigPubKey
-	fieldNumSigVal
-	fieldNumSigScheme
-)
-
-// Object message fields.
-const (
-	fieldNumObjectPayload = 4
-)
-
-// Replication request fields.
-const (
-	_ = iota
-	fieldNumReplicateObject
-	fieldNumReplicateSignature
-	fieldSignObjectMeta
-)
 
 type encodedObject struct {
 	b []byte
@@ -46,7 +27,7 @@ func encodeObjectWithoutPayload(hdr object.Object, pldLen int) (encodedObject, e
 
 	hdrProto := hdr.ProtoMessage()
 	hdrLen := hdrProto.MarshaledSize()
-	pldFldLen := protowire.SizeTag(fieldNumObjectPayload) + protowire.SizeBytes(pldLen)
+	pldFldLen := protowire.SizeTag(protoobject.FieldObjectPayload) + protowire.SizeBytes(pldLen)
 	if pldFldLen > math.MaxInt-hdrLen {
 		return res, fmt.Errorf("binary object is too big for this server: %d+%d>%d", hdrLen, pldLen, math.MaxInt)
 	}
@@ -60,7 +41,7 @@ func encodeObjectWithoutPayload(hdr object.Object, pldLen int) (encodedObject, e
 	res.b = res.b[:hdrLen]
 	hdrProto.MarshalStable(res.b)
 	res.pldFldOff = len(res.b)
-	res.b = protowire.AppendTag(res.b, fieldNumObjectPayload, protowire.BytesType)
+	res.b = protowire.AppendTag(res.b, protoobject.FieldObjectPayload, protowire.BytesType)
 	res.b = protowire.AppendVarint(res.b, uint64(pldLen))
 	res.pldOff = len(res.b)
 
@@ -81,7 +62,7 @@ func encodeReplicateRequestWithoutPayload(signer neofscrypto.Signer, hdr object.
 
 	hdrProto := hdr.ProtoMessage()
 	hdrLen := hdrProto.MarshaledSize()
-	pldFldLen := protowire.SizeTag(fieldNumObjectPayload) + protowire.SizeBytes(pldLen)
+	pldFldLen := protowire.SizeTag(protoobject.FieldObjectPayload) + protowire.SizeBytes(pldLen)
 	if pldFldLen > math.MaxInt-hdrLen {
 		return res, fmt.Errorf("binary object is too big for this server: %d+%d>%d", hdrLen, pldFldLen, math.MaxInt)
 	}
@@ -89,17 +70,17 @@ func encodeReplicateRequestWithoutPayload(signer neofscrypto.Signer, hdr object.
 	pubKey := neofscrypto.PublicKeyBytes(signer.Public())
 	sigScheme := signer.Scheme()
 
-	sigFldLen := protowire.SizeTag(fieldNumSigPubKey) + protowire.SizeBytes(len(pubKey)) +
-		protowire.SizeTag(fieldNumSigVal) + protowire.SizeBytes(len(sig)) +
-		protowire.SizeTag(fieldNumSigScheme) + protowire.SizeVarint(uint64(sigScheme))
-	fullLen := protowire.SizeTag(fieldNumReplicateSignature) + protowire.SizeBytes(sigFldLen) +
-		protowire.SizeTag(fieldNumReplicateObject)
+	sigFldLen := protowire.SizeTag(protorefs.FieldSignatureKey) + protowire.SizeBytes(len(pubKey)) +
+		protowire.SizeTag(protorefs.FieldSignatureValue) + protowire.SizeBytes(len(sig)) +
+		protowire.SizeTag(protorefs.FieldSignatureScheme) + protowire.SizeVarint(uint64(sigScheme))
+	fullLen := protowire.SizeTag(protoobject.FieldReplicateRequestSignature) + protowire.SizeBytes(sigFldLen) +
+		protowire.SizeTag(protoobject.FieldReplicateRequestObject)
 	objFldLen := hdrLen + pldFldLen
 	if protowire.SizeBytes(objFldLen) > math.MaxInt-fullLen {
 		return res, fmt.Errorf("replicate request exceeds server limit %d", math.MaxInt)
 	}
 	fullLen += protowire.SizeBytes(objFldLen)
-	fullLen += protowire.SizeTag(fieldSignObjectMeta) + protowire.SizeVarint(protowire.EncodeBool(signObjectMeta))
+	fullLen += protowire.SizeTag(protoobject.FieldReplicateRequestSignObject) + protowire.SizeVarint(protowire.EncodeBool(signObjectMeta))
 
 	res.b = getPayload()
 	if cap(res.b) < fullLen {
@@ -108,25 +89,25 @@ func encodeReplicateRequestWithoutPayload(signer neofscrypto.Signer, hdr object.
 	}
 
 	// meta signature extension flag
-	res.b = protowire.AppendTag(res.b, fieldSignObjectMeta, protowire.VarintType)
+	res.b = protowire.AppendTag(res.b, protoobject.FieldReplicateRequestSignObject, protowire.VarintType)
 	res.b = protowire.AppendVarint(res.b, protowire.EncodeBool(signObjectMeta))
 	// signature
-	res.b = protowire.AppendTag(res.b, fieldNumReplicateSignature, protowire.BytesType)
+	res.b = protowire.AppendTag(res.b, protoobject.FieldReplicateRequestSignature, protowire.BytesType)
 	res.b = protowire.AppendVarint(res.b, uint64(sigFldLen))
-	res.b = protowire.AppendTag(res.b, fieldNumSigPubKey, protowire.BytesType)
+	res.b = protowire.AppendTag(res.b, protorefs.FieldSignatureKey, protowire.BytesType)
 	res.b = protowire.AppendBytes(res.b, pubKey)
-	res.b = protowire.AppendTag(res.b, fieldNumSigVal, protowire.BytesType)
+	res.b = protowire.AppendTag(res.b, protorefs.FieldSignatureValue, protowire.BytesType)
 	res.b = protowire.AppendBytes(res.b, sig)
-	res.b = protowire.AppendTag(res.b, fieldNumSigScheme, protowire.VarintType)
+	res.b = protowire.AppendTag(res.b, protorefs.FieldSignatureScheme, protowire.VarintType)
 	res.b = protowire.AppendVarint(res.b, uint64(sigScheme))
 	// object
-	res.b = protowire.AppendTag(res.b, fieldNumReplicateObject, protowire.BytesType)
+	res.b = protowire.AppendTag(res.b, protoobject.FieldReplicateRequestObject, protowire.BytesType)
 	res.b = protowire.AppendVarint(res.b, uint64(objFldLen))
 	res.hdrOff = len(res.b)
 	res.b = res.b[:len(res.b)+hdrLen]
 	hdrProto.MarshalStable(res.b[res.hdrOff:])
 	res.pldFldOff = len(res.b)
-	res.b = protowire.AppendTag(res.b, fieldNumObjectPayload, protowire.BytesType)
+	res.b = protowire.AppendTag(res.b, protoobject.FieldObjectPayload, protowire.BytesType)
 	res.b = protowire.AppendVarint(res.b, uint64(pldLen))
 	res.pldOff = len(res.b)